+++ taglib/mpeg/id3v1/id3v1tag.h
@@ -140,7 +140,7 @@ namespace TagLib {
       /*!
        * Reads from the file specified in the constructor.
        */
-      void read(File *file, long tagOffset);
+      void read(File *file, long long tagOffset);
       /*!
        * Pareses the body of the tag in \a data.
        */
